Refactor code to consolidate utility imports
Removed redundant import statements and streamlined the way utility functions are imported in various files. This includes changes in 'Dump.js', 'actions.js', 'Installer.js', and 'Database.js'. The refactoring aims to improve code modularity and maintainability.
