+++ lib/Database.js
@@ -1,6 +1,6 @@
 const mysql = require('mysql2/promise');
-const { getConnectionSettings } = require('./data');
 const mysqldump = require('mysqldump');
+const {getConnectionSettings} = require("./utils/data");
 const fs = require('fs').promises;
 
 /**
