+++ lib/Dump.js
@@ -1,7 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const { getUserLocale } = require('./utils');
-const { getCurrentWpInfo, getLastWpVersion} = require("./utils/wpConfig");
+const { getCurrentWpInfo, getUserLocale, getLastWp} = require("./utils/wpConfig");
 
 /**
  * Represents a Dump class for WordPress configuration.
@@ -39,7 +38,7 @@ class Dump {
     const wpInfo = getCurrentWpInfo(this.wpFolder);
 
     const language = wpInfo.locale || getUserLocale();
-    const version = wpInfo.version || getLastWpVersion();
+    const version = wpInfo.version || getLastWp()?.version || 'latest';
 
     const result = {
       name,
