+++ lib/Installer.js
@@ -1,8 +1,8 @@
-const { isWPCLIAvailable } = require('./utils/data.js');
 const { cleanup, makeDir } = require('./utils/fs.js');
 const Package = require("./Package.js");
 const { WordPressPackage } = require('./WpPackage.js');
 const {runPostInstallCommands} = require("./utils/actions.js");
+const {isWPCLIAvailable} = require("./utils/actions");
 
 /**
  * The Installer class represents the WordPress installer and provides methods for installing WordPress and its dependencies.
