+++ lib/utils/actions.js
@@ -1,7 +1,6 @@
-import path from "path";
-import fs from "fs";
-import {child_process, exec} from "node:child_process";
-
+const path = require("node:path");
+const fs = require("fs");
+const {exec} = require("child_process");
 
 /**
  * Installs npm packages in the specified package directory.
@@ -78,7 +77,7 @@ async function runPostInstallCommands (commands) {
   for (const command of commands) {
     try {
       console.log(`Executing: ${command}`);
-      const { stdout, stderr } = await child_process.exec(command);
+      const { stdout, stderr } = await exec(command);
       if (stdout) {
         console.log(`Command output:\n${stdout}`);
       }
